healthbar is update correctly

+++ ETU/Game/Healthbar.h
@@ -10,10 +10,10 @@ public:
 
 	virtual bool init(const int maxHp);
 	virtual void draw(sf::RenderWindow& window) const;
-	void update(const int hp, sf::Vector2f characterPosition, float characterSize);
+	void update(const float hp, sf::Vector2f characterPosition, float characterSize);
 
 private:
-	int maxHp;
+	float maxHp;
 
 	sf::RectangleShape hpBar;
 	sf::RectangleShape hpBarBack;
